backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.10.1
dnspython==2.7.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-dotenv==1.1.1
httpx==0.28.1
pydantic-settings==2.10.1
